refaktorering: dedupliser toFailOrNull-metode, bedre logging i feilbehandler
