+++ feil-behandler/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/feilbehandler/river/FeilLytter.kt
@@ -1,6 +1,5 @@
 package no.nav.helsearbeidsgiver.inntektsmelding.feilbehandler.river
 
-import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.serializer
 import no.nav.helse.rapids_rivers.JsonMessage
 import no.nav.helse.rapids_rivers.MessageContext
@@ -10,6 +9,7 @@ import no.nav.helsearbeidsgiver.felles.BehovType
 import no.nav.helsearbeidsgiver.felles.Key
 import no.nav.helsearbeidsgiver.felles.json.toMap
 import no.nav.helsearbeidsgiver.felles.rapidsrivers.model.Fail
+import no.nav.helsearbeidsgiver.felles.rapidsrivers.model.ModelUtils.Companion.toFailOrNull
 import no.nav.helsearbeidsgiver.felles.rapidsrivers.toPretty
 import no.nav.helsearbeidsgiver.utils.json.fromJson
 import no.nav.helsearbeidsgiver.utils.json.parseJson
@@ -41,28 +41,24 @@ class FeilLytter(rapidsConnection: RapidsConnection) : River.PacketListener {
 
     override fun onPacket(packet: JsonMessage, context: MessageContext) {
         sikkerLogger.info("Mottok feil: ${packet.toPretty()}")
-        val fail = toFailOrNull(packet.toJson().parseJson())
-        if (fail == null) {
-            sikkerLogger.warn("Kunne ikke parse feil-objekt, ignorerer...")
-        } else if (skalHaandteres(fail)) {
-            sikkerLogger.info("Behandler feil")
-        } else {
-            sikkerLogger.info("Ignorerer feil")
+        val fail = toFailOrNull(packet.toJson().parseJson().toMap())
+        if (skalHaandteres(fail)) {
+            sikkerLogger.info("Behandler mottatt pakke")
         }
     }
 
-    fun skalHaandteres(fail: Fail): Boolean {
+    fun skalHaandteres(fail: Fail?): Boolean {
+        if (fail == null) {
+            sikkerLogger.warn("Kunne ikke parse feil-objekt, ignorerer...")
+            return false
+        }
         if (fail.forespoerselId == null) {
+            sikkerLogger.info("Mangler forespørselId, ignorerer")
             return false
         }
         val behovFraMelding = fail.utloesendeMelding.toMap()[Key.BEHOV]?.fromJson(BehovType.serializer())
-        return behovSomHaandteres.contains(behovFraMelding)
+        val skalHaandteres = behovSomHaandteres.contains(behovFraMelding)
+        sikkerLogger.info("Behov: $behovFraMelding skal håndteres: $skalHaandteres")
+        return skalHaandteres
     }
-
-    fun toFailOrNull(json: JsonElement): Fail? = // TODO: duplisert, lage felles metode?
-        json.toMap()[Key.FAIL]
-            ?.runCatching {
-                fromJson(Fail.serializer())
-            }
-            ?.getOrNull()
 }
